requirements.txt
PyQt6~=6.8.1
pywin32
PyVista~=0.44.1
pystray
pillow~=11.1.0
mss
vtk
numpy~=2.2.3
opencv-python-headless
scikit-image
